fix: 프로젝트 변경사항으로 인한 테이블 구조 변경
