+++ server/API/controller/room.ts
@@ -3,8 +3,8 @@ import { CustomRequest } from '../../type/type';
 interface RoomController {
   createRoom(req: Request, res: Response): void;
   closeRoom(req: Request, res: Response): void;
-  notifyUpdate(req: CustomRequest, res: Response): void;
-  updateRoom(req: CustomRequest, res: Response): void;
+  // notifyUpdate(req: CustomRequest, res: Response): void;
+  // updateRoom(req: CustomRequest, res: Response): void;
 }
 
 const roomController: RoomController = {
@@ -18,14 +18,14 @@ const roomController: RoomController = {
       message: 'ok',
     });
   },
-  notifyUpdate(req: CustomRequest, res: Response): void {
-    res.status(req.sendData.status).send({
-      message: req.sendData.message,
-    });
-  },
-  updateRoom(req: CustomRequest, res: Response): void {
-    // todo ...
-  },
+  // notifyUpdate(req: CustomRequest, res: Response): void {
+  //   res.status(req.sendData.status).send({
+  //     message: req.sendData.message,
+  //   });
+  // },
+  // updateRoom(req: CustomRequest, res: Response): void {
+  //   // todo ...
+  // },
 };
 
 export default roomController;
