+++ server/API/validation/room.ts
@@ -36,42 +36,41 @@ const roomValidation: RoomValidation = {
     res: Response,
     next: NextFunction
   ): Promise<void> {
-    const transaction: any = await db.sequelize.transaction();
-    let body: any = req.body;
-    if (
-      !body.id ||
-      !body.title ||
-      !body.allow_num ||
-      !body.category_id ||
-      !body.lon ||
-      !body.lat
-    ) {
-      res.status(400).send({
-        message: 'required data not received',
-      });
-      return;
-    }
-    let room = await db.Room.findOne({ where: { id: body.id } });
-    if (room) {
-      res.status(200).send({
-        message: 'room id aleady exists',
-      });
-      return;
-    }
-
-    if (body.is_private === 'Y')
-      body.password = await bcrypt.hashSync(body.password, 10);
-    await db.Room.create(body, { transaction });
-    await db.Participant.create(
-      {
-        room_id: body.id,
-        user_email: body.user_email,
-        role: 'HOST',
-      },
-      { transaction }
-    );
-    await transaction.commit();
-    next();
+    // const transaction: any = await db.sequelize.transaction();
+    // let body: any = req.body;
+    // if (
+    //   !body.id ||
+    //   !body.title ||
+    //   !body.allow_num ||
+    //   !body.category_id ||
+    //   !body.lon ||
+    //   !body.lat
+    // ) {
+    //   res.status(400).send({
+    //     message: 'required data not received',
+    //   });
+    //   return;
+    // }
+    // let room = await db.Room.findOne({ where: { id: body.id } });
+    // if (room) {
+    //   res.status(200).send({
+    //     message: 'room id aleady exists',
+    //   });
+    //   return;
+    // }
+    // if (body.is_private === 'Y')
+    //   body.password = await bcrypt.hashSync(body.password, 10);
+    // await db.Room.create(body, { transaction });
+    // await db.Participant.create(
+    //   {
+    //     room_id: body.id,
+    //     user_email: body.user_email,
+    //     role: 'HOST',
+    //   },
+    //   { transaction }
+    // );
+    // await transaction.commit();
+    // next();
   },
 
   /*
@@ -82,23 +81,23 @@ const roomValidation: RoomValidation = {
     res: Response,
     next: NextFunction
   ): Promise<void> {
-    let id: string = req.body.room_id;
-    let find: any = await db.Room.findOne({
-      attributes: {
-        exclude: [
-          /*'UserId', 'CategoryId'*/
-        ],
-      },
-      where: { id },
-    });
-    if (!find) {
-      res.status(409).send({
-        message: 'roomId not exist',
-      });
-      return;
-    }
-    await db.Room.update({ is_close: 'Y' }, { where: { id } });
-    next();
+    // let id: string = req.body.room_id;
+    // let find: any = await db.Room.findOne({
+    //   attributes: {
+    //     exclude: [
+    //       /*'UserId', 'CategoryId'*/
+    //     ],
+    //   },
+    //   where: { id },
+    // });
+    // if (!find) {
+    //   res.status(409).send({
+    //     message: 'roomId not exist',
+    //   });
+    //   return;
+    // }
+    // await db.Room.update({ is_close: 'Y' }, { where: { id } });
+    // next();
   },
   /*
     모임 공지 생성
@@ -108,22 +107,21 @@ const roomValidation: RoomValidation = {
     res: Response,
     next: NextFunction
   ): Promise<void> {
-    let body = req.body;
-    if (!body.notify || body.notify === '') {
-      req.sendData = { message: 'notification not received', status: 400 };
-      next();
-      return;
-    }
-    let find = await db.Room.findOne({ where: { id: body.id } });
-    if (!find) {
-      req.sendData = { message: 'room not exist', status: 401 };
-      next();
-      return;
-    }
-
-    await db.Room.update({ notify: body.notify }, { where: { id: body.id } });
-    req.sendData = { message: 'ok', status: 200 };
-    next();
+    // let body = req.body;
+    // if (!body.notify || body.notify === '') {
+    //   req.sendData = { message: 'notification not received', status: 400 };
+    //   next();
+    //   return;
+    // }
+    // let find = await db.Room.findOne({ where: { id: body.id } });
+    // if (!find) {
+    //   req.sendData = { message: 'room not exist', status: 401 };
+    //   next();
+    //   return;
+    // }
+    // await db.Room.update({ notify: body.notify }, { where: { id: body.id } });
+    // req.sendData = { message: 'ok', status: 200 };
+    // next();
   },
   /*
     모임 업데이트
@@ -133,7 +131,24 @@ const roomValidation: RoomValidation = {
     res: Response,
     next: NextFunction
   ): Promise<void> {
-    // todo ...
+    // let body: any = req.body;
+    // let id: string = body.id;
+    // let count: any = await db.Participant.findOne({
+    //   attributes: [[db.sequelize.fn('COUNT', 'id'), 'num_count']],
+    //   group: ['room_id'],
+    //   where: { room_id: body.id },
+    // });
+    // if (count && count.dataValues.num_count > body.allow_num) {
+    //   req.sendData = { message: 'number range exit', status: 409 };
+    //   next();
+    //   return;
+    // }
+    // if (body.is_private === 'Y')
+    //   body.password = await bcrypt.hashSync(body.password, 10);
+    // delete body.id;
+    // await db.Room.update(body, { where: { id } });
+    // req.sendData = { message: 'ok', status: 200 };
+    // next();
   },
 };
